improving engineering toolkit structure

+++ Week01/EngineeringToolkit/calculator.h
@@ -1,5 +1,5 @@
-int add(int a, int b);
-int subtract(int a, int b);
-int multiply(int a, int b);
+float add(float a, float b);
+float subtract(float a, float b);
+float multiply(float a, float b);
 float divide(float a, float b);
 void calculator(void);
